admin: split out security endpoints
