+++ src/v/redpanda/admin/server.cc
@@ -186,15 +186,6 @@ private:
     ss::lw_shared_ptr<C> c_;
 };
 
-struct string_conversion_exception : public default_control_character_thrower {
-    using default_control_character_thrower::default_control_character_thrower;
-    [[noreturn]] [[gnu::cold]] void conversion_error() override {
-        throw ss::httpd::bad_request_exception(
-          "Parameter contained invalid control characters: "
-          + get_sanitized_string());
-    }
-};
-
 inline net::unresolved_address from_ss_sa(const ss::socket_address& sa) {
     return {fmt::format("{}", sa.addr()), sa.port(), sa.addr().in_family()};
 }
@@ -939,8 +930,7 @@ ss::future<ss::httpd::redirect_exception> admin_server::redirect_to_leader(
       url, ss::http::reply::status_type::temporary_redirect);
 }
 
-namespace {
-bool need_redirect_to_leader(
+bool admin_server::need_redirect_to_leader(
   model::ntp ntp, ss::sharded<cluster::metadata_cache>& metadata_cache) {
     auto leader_id_opt = metadata_cache.local().get_leader_id(ntp);
     if (!leader_id_opt.has_value()) {
@@ -952,7 +942,6 @@ bool need_redirect_to_leader(
 
     return leader_id_opt.value() != *config::node().node_id();
 }
-} // namespace
 
 model::node_id admin_server::parse_broker_id(const ss::http::request& req) {
     try {
@@ -1962,281 +1951,6 @@ void admin_server::register_raft_routes() {
       ss::httpd::raft_json::get_raft_recovery_status,
       [this](std::unique_ptr<ss::http::request> req) {
           return get_raft_recovery_status_handler(std::move(req));
-      });
-}
-
-namespace {
-
-// TODO: factor out generic serialization from seastar http exceptions
-security::scram_credential parse_scram_credential(const json::Document& doc) {
-    if (!doc.IsObject()) {
-        throw ss::httpd::bad_request_exception(fmt::format("Not an object"));
-    }
-
-    if (!doc.HasMember("algorithm") || !doc["algorithm"].IsString()) {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("String algo missing"));
-    }
-    const auto algorithm = std::string_view(
-      doc["algorithm"].GetString(), doc["algorithm"].GetStringLength());
-    validate_no_control(algorithm, string_conversion_exception{algorithm});
-
-    if (!doc.HasMember("password") || !doc["password"].IsString()) {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("String password smissing"));
-    }
-    const auto password = doc["password"].GetString();
-    validate_no_control(password, string_conversion_exception{"PASSWORD"});
-
-    security::scram_credential credential;
-
-    if (algorithm == security::scram_sha256_authenticator::name) {
-        credential = security::scram_sha256::make_credentials(
-          password, security::scram_sha256::min_iterations);
-
-    } else if (algorithm == security::scram_sha512_authenticator::name) {
-        credential = security::scram_sha512::make_credentials(
-          password, security::scram_sha512::min_iterations);
-
-    } else {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("Unknown scram algorithm: {}", algorithm));
-    }
-
-    return credential;
-}
-
-bool match_scram_credential(
-  const json::Document& doc, const security::scram_credential& creds) {
-    // Document is pre-validated via earlier parse_scram_credential call
-    const auto password = ss::sstring(doc["password"].GetString());
-    const auto algorithm = std::string_view(
-      doc["algorithm"].GetString(), doc["algorithm"].GetStringLength());
-    validate_no_control(algorithm, string_conversion_exception{algorithm});
-
-    if (algorithm == security::scram_sha256_authenticator::name) {
-        return security::scram_sha256::validate_password(
-          password, creds.stored_key(), creds.salt(), creds.iterations());
-    } else if (algorithm == security::scram_sha512_authenticator::name) {
-        return security::scram_sha512::validate_password(
-          password, creds.stored_key(), creds.salt(), creds.iterations());
-    } else {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("Unknown scram algorithm: {}", algorithm));
-    }
-}
-
-bool is_no_op_user_write(
-  security::credential_store& store,
-  security::credential_user username,
-  security::scram_credential credential) {
-    auto user_opt = store.get<security::scram_credential>(username);
-    if (user_opt.has_value()) {
-        return user_opt.value() == credential;
-    } else {
-        return false;
-    }
-}
-} // namespace
-
-ss::future<ss::json::json_return_type>
-admin_server::create_user_handler(std::unique_ptr<ss::http::request> req) {
-    if (need_redirect_to_leader(model::controller_ntp, _metadata_cache)) {
-        // In order that we can do a reliably ordered validation of
-        // the request (and drop no-op requests), run on controller leader;
-        throw co_await redirect_to_leader(*req, model::controller_ntp);
-    }
-
-    auto doc = co_await parse_json_body(req.get());
-
-    auto credential = parse_scram_credential(doc);
-
-    if (!doc.HasMember("username") || !doc["username"].IsString()) {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("String username missing"));
-    }
-
-    auto username = security::credential_user(doc["username"].GetString());
-    validate_no_control(username(), string_conversion_exception{username()});
-
-    if (is_no_op_user_write(
-          _controller->get_credential_store().local(), username, credential)) {
-        vlog(
-          adminlog.debug,
-          "User {} already exists with matching credential",
-          username);
-        co_return ss::json::json_return_type(ss::json::json_void());
-    }
-
-    auto err
-      = co_await _controller->get_security_frontend().local().create_user(
-        username, credential, model::timeout_clock::now() + 5s);
-    vlog(
-      adminlog.debug, "Creating user '{}' {}:{}", username, err, err.message());
-
-    if (err == cluster::errc::user_exists) {
-        // Idempotency: if user is same as one that already exists,
-        // suppress the user_exists error and return success.
-        const auto& credentials_store
-          = _controller->get_credential_store().local();
-        std::optional<security::scram_credential> creds
-          = credentials_store.get<security::scram_credential>(username);
-        if (creds.has_value() && match_scram_credential(doc, creds.value())) {
-            co_return ss::json::json_return_type(ss::json::json_void());
-        }
-    }
-
-    co_await throw_on_error(*req, err, model::controller_ntp);
-    co_return ss::json::json_return_type(ss::json::json_void());
-}
-
-ss::future<ss::json::json_return_type>
-admin_server::delete_user_handler(std::unique_ptr<ss::http::request> req) {
-    if (need_redirect_to_leader(model::controller_ntp, _metadata_cache)) {
-        // In order that we can do a reliably ordered validation of
-        // the request (and drop no-op requests), run on controller leader;
-        throw co_await redirect_to_leader(*req, model::controller_ntp);
-    }
-
-    auto user = security::credential_user(req->param["user"]);
-
-    if (!_controller->get_credential_store().local().contains(user)) {
-        vlog(adminlog.debug, "User '{}' already gone during deletion", user);
-        co_return ss::json::json_return_type(ss::json::json_void());
-    }
-
-    auto err
-      = co_await _controller->get_security_frontend().local().delete_user(
-        user, model::timeout_clock::now() + 5s);
-    vlog(adminlog.debug, "Deleting user '{}' {}:{}", user, err, err.message());
-    if (err == cluster::errc::user_does_not_exist) {
-        // Idempotency: removing a non-existent user is successful.
-        co_return ss::json::json_return_type(ss::json::json_void());
-    }
-    co_await throw_on_error(*req, err, model::controller_ntp);
-    co_return ss::json::json_return_type(ss::json::json_void());
-}
-
-ss::future<ss::json::json_return_type>
-admin_server::update_user_handler(std::unique_ptr<ss::http::request> req) {
-    if (need_redirect_to_leader(model::controller_ntp, _metadata_cache)) {
-        // In order that we can do a reliably ordered validation of
-        // the request (and drop no-op requests), run on controller leader;
-        throw co_await redirect_to_leader(*req, model::controller_ntp);
-    }
-
-    auto user = security::credential_user(req->param["user"]);
-
-    auto doc = co_await parse_json_body(req.get());
-
-    auto credential = parse_scram_credential(doc);
-
-    if (is_no_op_user_write(
-          _controller->get_credential_store().local(), user, credential)) {
-        vlog(
-          adminlog.debug,
-          "User {} already exists with matching credential",
-          user);
-        co_return ss::json::json_return_type(ss::json::json_void());
-    }
-
-    auto err
-      = co_await _controller->get_security_frontend().local().update_user(
-        user, credential, model::timeout_clock::now() + 5s);
-    vlog(adminlog.debug, "Updating user {}:{}", err, err.message());
-    co_await throw_on_error(*req, err, model::controller_ntp);
-    co_return ss::json::json_return_type(ss::json::json_void());
-}
-
-ss::future<ss::json::json_return_type>
-admin_server::oidc_whoami_handler(std::unique_ptr<ss::http::request> req) {
-    auto auth_hdr = req->get_header("authorization");
-    if (!auth_hdr.starts_with(authz_bearer_prefix)) {
-        throw ss::httpd::base_exception{
-          "Invalid Authorization header",
-          ss::http::reply::status_type::unauthorized};
-    }
-
-    security::oidc::authenticator auth{_controller->get_oidc_service().local()};
-    auto res = auth.authenticate(auth_hdr.substr(authz_bearer_prefix.length()));
-
-    if (res.has_error()) {
-        throw ss::httpd::base_exception{
-          "Invalid Authorization header",
-          ss::http::reply::status_type::unauthorized};
-    }
-
-    ss::httpd::security_json::oidc_whoami_response j_res{};
-    j_res.id = res.assume_value().principal.name();
-    j_res.expire = res.assume_value().expiry.time_since_epoch() / 1s;
-
-    co_return ss::json::json_return_type(j_res);
-}
-
-ss::future<ss::json::json_return_type>
-admin_server::oidc_keys_cache_invalidate_handler(
-  std::unique_ptr<ss::http::request> req) {
-    auto f = co_await ss::coroutine::as_future(
-      _controller->get_oidc_service().invoke_on_all(
-        [](auto& s) { return s.refresh_keys(); }));
-    if (f.failed()) {
-        ss::httpd::security_json::oidc_keys_cache_invalidate_error_response res;
-        res.error_message = ssx::sformat("", f.get_exception());
-        co_return ss::json::json_return_type(res);
-    }
-    co_return ss::json::json_return_type(ss::json::json_void());
-}
-
-void admin_server::register_security_routes() {
-    register_route<superuser>(
-      ss::httpd::security_json::create_user,
-      [this](std::unique_ptr<ss::http::request> req) {
-          return create_user_handler(std::move(req));
-      });
-
-    register_route<superuser>(
-      ss::httpd::security_json::delete_user,
-      [this](std::unique_ptr<ss::http::request> req) {
-          return delete_user_handler(std::move(req));
-      });
-
-    register_route<superuser>(
-      ss::httpd::security_json::update_user,
-      [this](std::unique_ptr<ss::http::request> req) {
-          return update_user_handler(std::move(req));
-      });
-
-    register_route<user>(
-      ss::httpd::security_json::oidc_whoami,
-      [this](std::unique_ptr<ss::http::request> req) {
-          return oidc_whoami_handler(std::move(req));
-      });
-
-    register_route<superuser>(
-      ss::httpd::security_json::oidc_keys_cache_invalidate,
-      [this](std::unique_ptr<ss::http::request> req) {
-          return oidc_keys_cache_invalidate_handler(std::move(req));
-      });
-
-    register_route<superuser>(
-      ss::httpd::security_json::list_users,
-      [this](std::unique_ptr<ss::http::request> req) {
-          bool include_ephemeral = req->get_query_param("include_ephemeral")
-                                   == "true";
-
-          auto pred = [include_ephemeral](auto const& c) {
-              return include_ephemeral
-                     || security::credential_store::is_not_ephemeral(c);
-          };
-          auto creds = _controller->get_credential_store().local().range(pred);
-
-          std::vector<ss::sstring> users{};
-          users.reserve(std::distance(creds.begin(), creds.end()));
-          for (const auto& [user, type] : creds) {
-              users.push_back(user());
-          }
-          return ss::make_ready_future<ss::json::json_return_type>(
-            std::move(users));
       });
 }
